fix training script
